Gather three functions into one module
